--- src/common/components/FormSubTitle/index.ts
@@ -1,2 +1,0 @@
-export { FormSubTitle } from "./FormSubTitle";
-export type { FormSubTitleProps } from "./FormSubTitle";
