--- src/domains/CompanySignup/index.ts
@@ -1,1 +1,0 @@
-export { CompanySignUpView } from "./components/CompanySingupView";
