+++ src/domains/Signup/BasicInfo/components/TermsForm/AgreeAllButton/AgreeAllButton.tsx
