+++ src/domains/Signup/BasicInfo/components/TermsForm/AgreeAllButton/index.ts
