+++ src/domains/Signup/BasicInfo/components/TermsForm/TermItemRow/TermItemRow.tsx
