+++ src/domains/Signup/BasicInfo/components/TermsForm/TermItemRow/index.ts
