+++ src/domains/Signup/BasicInfo/components/TermsForm/TermsForm.tsx
