+++ src/domains/Signup/BasicInfo/components/TermsForm/index.ts
