+++ src/domains/Signup/CompanyInfo/components/SignupCompanyInfoView/index.ts
@@ -1,0 +1,1 @@
+export { SignupCompanyInfoView } from "./SignupCompanyInfoView";
