+++ src/domains/Signup/CompanyInfo/index.ts
@@ -1,0 +1,1 @@
+export { SignupCompanyInfoView } from "./components/SignupCompanyInfoView";
