+++ src/domains/Signup/common/components/FieldHeader/index.ts
@@ -1,0 +1,2 @@
+export { FieldHeader } from "./FieldHeader";
+export type { FieldHeaderProps } from "./FieldHeader";
