+++ src/domains/Signup/common/components/FormTitle/FormTitle.tsx
@@ -1,0 +1,5 @@
+import { Typography, TypographyProps } from "@mui/material";
+
+export const FormTitle = (p: TypographyProps) => (
+  <Typography variant="h6" fontWeight={"400"} {...p} />
+);
