+++ src/domains/Signup/common/components/FormTitle/index.ts
@@ -1,0 +1,1 @@
+export { FormTitle } from "./FormTitle";
