+++ src/domains/Signup/common/components/SmallCaption/index.ts
@@ -1,0 +1,1 @@
+export { SmallCaption } from "./SmallCaption";
