+++ src/domains/Signup/common/components/templates/FieldHeaderWithContent/index.ts
@@ -1,0 +1,1 @@
+export { FieldHeaderWithContent } from "./FieldHeaderWithContent";
