--- src/domains/Signup/components/BasicInfo/components/FieldHeader/index.ts
@@ -1,1 +1,0 @@
-export { FieldHeader } from "./FieldHeader";
