+++ src/domains/Signup/index.ts
@@ -1,1 +1,2 @@
-export { SignupBasicInfoView } from "./components/BasicInfo";
+export { SignupBasicInfoView } from "./BasicInfo";
+export { SignupCompanyInfoView } from "./CompanyInfo";
